+++ src/main/java/com/ms/ms_bets/infra/adapter/repository/BetRepositoryAccess.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BetRepositoryAccess extends JpaRepository<BetModel, Integer> {
+public interface BetRepositoryAccess extends JpaRepository<BetModel, Long> {
 
     List<BetModel> findAllByEventId(Long eventId);
